requirements.txt
Django>=5.0,<6.0           # Web framework
djangorestframework>=3.14  # REST API layer
mysqlclient>=2.2           # MySQL database driver
django-environ>=0.11       # Load environment variables from .env